Improve simulation performance
The forecast thing was very inefficent .. there is no point in getting
*ALL* results from the database and then only use one ... limit on the
db side instead. This gives a huge boost (from takes forever to takes
a bit over 2 minutes).

In addition to that we only need the time and temperature fields so use
a project and limit to those fields only. This change cuts the run time
in half again.

Also add a runtime measurement to the simulator.